Create woo-ocrpdf.py
Launched with as a parameter a PDF file or a folder containing PDF files, this python script will evaluate if a PDF is searchable and if not OCR it and create a searchable PDF.
The non-searchable PDF will be copied into a sub-folder of the folder containing the PDF with "_ns" added to the name. The newly created PDF will overwrite the non-searchable PDF in the original folder.